updated 5

+++ src/main/java/ge/softlab/homework/store_levan_demo/model/Product.java
@@ -26,7 +26,7 @@ public class Product {
     @Column(name = "purchase_price")
     private Integer purchasePrice;
     @Column(name = "sell_price")
-    private Integer sellPrice;
+    private Double sellPrice;
     @Column(name = "remaining")
     private Integer remaining;
